Fix compile on Windows
TBR=yurys@chromium.org,dslomov@chromium.org
LOG=N

Review URL: https://codereview.chromium.org/170543002

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@19454 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/heap-snapshot-generator.cc
@@ -908,7 +908,8 @@ HeapEntry* V8HeapExplorer::AddEntry(Address address,
                                     HeapEntry::Type type,
                                     const char* name,
                                     size_t size) {
-  SnapshotObjectId object_id = heap_object_map_->FindOrAddEntry(address, size);
+  SnapshotObjectId object_id = heap_object_map_->FindOrAddEntry(
+      address, static_cast<unsigned int>(size));
   return snapshot_->AddEntry(type, name, object_id, size);
 }
